+++ tagpack/tagpack.py
@@ -1,14 +1,107 @@
 """TagPack - A wrapper for TagPacks files"""
+import glob
 import json
 import os
 import sys
 import yaml
-
+from git import Repo
+import giturlparse as gup
+import pathlib
 from tagpack import TagPackFileError, ValidationError
+from yamlinclude import YamlIncludeConstructor
+
+from tagpack.cmd_utils import print_info
+
+
+def get_repository(path: str) -> pathlib.Path:
+    """ Parse '/home/anna/graphsense/graphsense-tagpacks/public/packs' ->
+        and return pathlib object with the repository root
+     """
+    repo_path = pathlib.Path(path)
+
+    while str(repo_path) != repo_path.root:
+        try:
+            Repo(repo_path)
+            return repo_path
+        except:
+            pass
+        repo_path = repo_path.parent
+    raise ValidationError(f'No repository root found in path {path}')
+
+
+def get_uri_for_tagpack(repo_path, tagpack_file, strict_check, no_git):
+    """ For a given path string
+    '/home/anna/graphsense/graphsense-tagpacks/public/packs'
+    and tagpack file string
+    '/home/anna/graphsense/graphsense-tagpacks/public/packs/a/2021/01/2010101.yaml'
+
+    return remote URI
+        'https://github.com/anna/tagpacks/blob/develop/public/packs/a/2021/01/2010101.yaml'
+    or, if no_git is set
+        tagpack file string
+
+    Local git copy will be checked for modifications by default.
+    Toggle strict_check param to change this.
+
+    If path does not contain git information, the original path
+    is returned.
+    """
+    if no_git:
+        return tagpack_file
+
+    repo = Repo(repo_path)
+
+    if strict_check and repo.is_dirty():
+        print_info(f"Local modifications in {repo.common_dir} detected, please push first.")
+        sys.exit(0)
+
+    if len(repo.remotes) > 1:
+        raise ValidationError("Multiple remotes present, cannot decide on backlink. ")
+
+    rel_path = pathlib.Path(tagpack_file).relative_to(repo_path)
+
+    u = next(repo.remotes[0].urls)
+    g = gup.parse(u).url2https.replace('.git', '')
+    res = f'{g}/tree/{repo.active_branch.name}/{rel_path}'
+    return res
+
+
+def collect_tagpack_files(path):
+    """Collect Tagpack YAML files from given paths"""
+    tagpack_files = []
+    header_path = None
+
+    if os.path.isdir(path):
+        files = glob.glob(path + '/**/*.yaml', recursive=True)
+        tagpack_files = tagpack_files + files
+    elif os.path.isfile(path):  # validate single file
+        tagpack_files.append(path)
+
+    # check if corresponding yaml header file exists in a parent directory
+    header_dir = path
+    while header_dir and header_dir != os.path.sep:
+        header_dir, _ = os.path.split(header_dir)
+        res = glob.glob(os.path.join(header_dir, 'header.yaml'))
+        if res:
+            tagpack_files += res
+            break
+
+    # deal with yaml includes
+    tagpack_files = set(tagpack_files)
+    for p in tagpack_files:
+        if p.endswith('header.yaml'):
+            header_path = p
+    if header_path:
+        tagpack_files.remove(header_path)
+        header_path = os.path.dirname(header_path)
+
+    tagpack_files = [a for a in tagpack_files if not a.endswith('config.yaml')]
+
+    return tagpack_files, header_path
 
 
 # https://gist.github.com/pypt/94d747fe5180851196eb
-class UniqueKeyLoader(yaml.SafeLoader):
+class UniqueKeyLoader(yaml.FullLoader):
     def construct_mapping(self, node, deep=False):
         mapping = set()
         for key_node, value_node in node.value:
@@ -27,15 +120,21 @@ class TagPack(object):
         self.contents = contents
         self.schema = schema
         self.taxonomies = taxonomies
+        self._unique_tags = []
+        self._duplicates = []
 
-    def load_from_file(baseuri, pathname, schema, taxonomies):
+    def load_from_file(uri, pathname, schema, taxonomies, header_dir=None):
+        YamlIncludeConstructor.add_to_loader_class(loader_class=yaml.FullLoader, base_dir=header_dir)
+
         if not os.path.isfile(pathname):
             sys.exit("This program requires {} to be a file"
                      .format(pathname))
         contents = yaml.load(open(pathname, 'r'), UniqueKeyLoader)
-        if not baseuri.endswith(os.path.sep):
-            baseuri = baseuri + os.path.sep
-        uri = baseuri + pathname
+
+        if 'header' in contents.keys():
+            for k, v in contents['header'].items():
+                contents[k] = v
+            contents.pop('header')
         return TagPack(uri, contents, schema, taxonomies)
 
     @property
@@ -73,6 +172,26 @@ class TagPack(object):
         except AttributeError:
             raise TagPackFileError("Cannot extract tags from tagpack")
 
+    def get_unique_tags(self):
+        if self._unique_tags:
+            return self._unique_tags
+
+        seen = set()
+        duplicates = []
+
+        for tag in self.tags:
+            # check if duplicate entry
+            t = tuple([str(tag.all_fields.get(k)).lower() if k in tag.all_fields.keys() else ''
+                       for k in ['address', 'currency', 'label', 'source']])
+            if t in seen:
+                duplicates.append(t)
+            else:
+                seen.add(t)
+                self._unique_tags.append(tag)
+
+        self._duplicates = duplicates
+        return self._unique_tags
+
     def validate(self):
         """Validates a TagPack against its schema and used taxonomies"""
 
@@ -97,16 +216,11 @@ class TagPack(object):
             self.schema.check_type(field, value)
             self.schema.check_taxonomies(field, value, self.taxonomies)
 
-        # iterate over all tags, report duplicates, check types, taxonomy and mandatory use
-        seen = set()
+        if len(self.tags) < 1:
+            raise ValidationError("no tags found.")
 
-        for tag in self.tags:
-            # check if duplicate entry
-            t = tuple([tag.all_fields.get(k) for k in ['address', 'currency', 'label', 'source']])
-            if t in seen:
-                raise ValidationError("Duplicate found {}".format(t))
-            seen.add(t)
-
+        # iterate over all tags, check types, taxonomy and mandatory use
+        for tag in self.get_unique_tags():
             # check if mandatory tag fields are defined
             if not isinstance(tag, Tag):
                 raise ValidationError("Unknown tag type {}".format(tag))
@@ -114,25 +228,27 @@ class TagPack(object):
             for schema_field in self.schema.mandatory_tag_fields:
                 if schema_field not in tag.explicit_fields and \
                    schema_field not in self.tag_fields:
-                    raise ValidationError("Mandatory tag field {} missing"
-                                          .format(schema_field))
+                    raise ValidationError(f"Mandatory tag field {schema_field} missing in {tag} ")
 
             for field, value in tag.explicit_fields.items():
                 # check whether field is defined as body field
                 if field not in self.schema.tag_fields:
-                    raise ValidationError("Field {} not allowed in tag"
-                                          .format(field))
+                    raise ValidationError(f"Field {field} not allowed in {tag} ")
 
                 # check for None values
                 if value is None:
                     raise ValidationError(
-                        "Value of body field {} must not be empty (None)"
-                        .format(field))
+                        f"Value of body field {field} must not be empty (None) in {tag}")
 
                 # check types and taxomomy use
-                self.schema.check_type(field, value)
-                self.schema.check_taxonomies(field, value, self.taxonomies)
+                try:
+                    self.schema.check_type(field, value)
+                    self.schema.check_taxonomies(field, value, self.taxonomies)
+                except ValidationError as e:
+                    raise ValidationError(f'{e} in {tag}')
 
+        if self._duplicates:
+            print_info(f"{len(self._duplicates)} duplicate(s) found, starting with {self._duplicates[0]}\n")
         return True
 
     def to_json(self):
@@ -168,7 +284,7 @@ class Tag(object):
     @property
     def all_fields(self):
         """Return all tag fields (explicit and generic)"""
-        return {**self.explicit_fields, **self.tagpack.tag_fields}
+        return {**self.tagpack.tag_fields, **self.explicit_fields, }
 
     def to_json(self):
         """Returns a JSON serialization of all tag fields"""