+++ tests/test_tagpack.py
@@ -1,9 +1,10 @@
 import json
 from datetime import date
+
 import pytest
 
-from tagpack import TagPackFileError, ValidationError
-from tagpack.tagpack import TagPack, Tag
+from tagpack import ValidationError
+from tagpack.tagpack import TagPack, Tag, collect_tagpack_files
 from tagpack.tagpack_schema import TagPackSchema
 from tagpack.taxonomy import Taxonomy
 
@@ -38,14 +39,59 @@ def tagpack(schema, taxonomies):
                     'tags': [
                         {'label': 'Some attribution tag',
                          'address': '123Bitcoin45'
-                         },
+                         },  # inherits all header fields
                         {'label': 'Another attribution tag',
                          'address': '123Bitcoin66',
-                         'context': '{"counts": 1}'
-                         }
+                         'context': '{"counts": 1}',
+                         'currency': 'ETH',
+                         }  # overrides currency
                     ]},
                    schema,
                    taxonomies)
+
+
+def test_all_header_fields(tagpack, schema):
+    h = ['title', 'creator', 'source', 'currency', 'lastmod', 'tags']
+
+    assert all(field in tagpack.all_header_fields
+               for field in
+               h)
+    assert len(h) == len(tagpack.all_header_fields)
+
+    assert tagpack.all_header_fields.get('creator') == 'GraphSense Team'
+    assert tagpack.all_header_fields.get('title') == 'Test TagPack'
+
+
+def test_header_fields(tagpack):
+    h = ['title', 'creator', 'tags']
+
+    assert all(field in tagpack.header_fields
+               for field in h)
+    assert len(h) == len(tagpack.header_fields)
+
+
+def test_explicit_tag_fields(tagpack):
+    tag = tagpack.tags[0]
+    t = ['label', 'address']
+    assert all(field in tag.explicit_fields for field in t)
+    assert len(t) == len(tag.explicit_fields)
+
+    tag = tagpack.tags[1]
+    t = ['label', 'address', 'context', 'currency']
+    assert all(field in tag.explicit_fields for field in t)
+    assert len(t) == len(tag.explicit_fields)
+
+
+def test_tag_inherits_from_header(tagpack):
+    t = tagpack.tags[0]
+
+    assert t.all_fields.get('currency') == 'BTC'
+
+
+def test_tag_overrides_header(tagpack):
+    t = tagpack.tags[1]
+
+    assert t.all_fields.get('currency') == 'ETH'
 
 
 def test_init(tagpack):
@@ -55,16 +101,6 @@ def test_init(tagpack):
     assert isinstance(tagpack.schema, TagPackSchema)
     assert 'title' in tagpack.schema.header_fields
     assert isinstance(tagpack.taxonomies, dict)
-
-
-def test_all_header_fields(tagpack, schema):
-    assert all(field in tagpack.all_header_fields
-               for field in ['title', 'creator', 'lastmod', 'tags'])
-
-
-def test_header_fields(tagpack):
-    assert all(field in tagpack.header_fields
-               for field in ['title', 'creator'])
 
 
 def test_tag_fields(tagpack):
@@ -245,22 +281,107 @@ def test_validate_fail_empty_body_field(tagpack):
         in str(e.value)
 
 
-def test_duplicate_raises(tagpack):
-    tagpack.contents['tags'].append(tagpack.contents['tags'][0])
-    with pytest.raises(ValidationError) as e:
-        tagpack.validate()
-    assert "Duplicate found" in str(e.value)
+def test_simple_file_collection():
+    prefix = 'tests/testfiles/simple/'
+    files, headerfile_path = collect_tagpack_files(prefix)
+
+    assert len(files) == 4
+    assert f'{prefix}ex_addr_tagpack.yaml' in files
+    assert f'{prefix}duplicate_tag.yaml' in files
+    assert f'{prefix}empty_tag_list.yaml' in files
+    assert f'{prefix}multiple_tags_for_address.yaml' in files
 
 
-def test_load_from_file_addr_tagpack(schema, taxonomies):
-    tagpack = TagPack.load_from_file('http://example.com/',
-                                     'tests/testfiles/ex_addr_tagpack.yaml',
-                                     schema,
+def test_file_collection_with_yaml_include():
+    files, headerfile_path = collect_tagpack_files('tests/testfiles/yaml_inclusion/')
+
+    assert len(files) == 4
+    assert 'tests/testfiles/yaml_inclusion/2021/01/20210101.yaml' in files
+    assert 'tests/testfiles/yaml_inclusion/2021/01/20210102.yaml' in files
+    assert 'tests/testfiles/yaml_inclusion/2021/02/20210201.yaml' in files
+    assert 'tests/testfiles/yaml_inclusion/2021/01/special/20210106-special.yaml' in files
+
+    assert headerfile_path == 'tests/testfiles/yaml_inclusion'
+
+
+def test_file_collection_with_missing_yaml_include_raises_exception():
+    files, headerfile_path = collect_tagpack_files('tests/testfiles/yaml_inclusion_missing_header/')
+
+    assert not headerfile_path
+
+    with pytest.raises(FileNotFoundError) as e:
+        TagPack.load_from_file(None, files[0], None, None, headerfile_path)
+    assert "No such file or directory: 'header.yaml'" \
+        in str(e.value)
+
+
+def test_load_from_file_addr_tagpack(taxonomies):
+    tagpack = TagPack.load_from_file('http://example.com/packs/ex_addr_tagpack.yaml',
+                                     'tests/testfiles/simple/ex_addr_tagpack.yaml',
+                                     TagPackSchema(),
                                      taxonomies)
     assert isinstance(tagpack, TagPack)
     assert tagpack.uri == \
-        'http://example.com/tests/testfiles/ex_addr_tagpack.yaml'
+        'http://example.com/packs/ex_addr_tagpack.yaml'
     assert tagpack.contents['title'] == 'Test Address TagPack'
     assert isinstance(tagpack.schema, TagPackSchema)
     assert 'title' in tagpack.schema.header_fields
     assert isinstance(tagpack.taxonomies, dict)
+
+
+def test_yaml_inclusion(taxonomies):
+    tagpack = TagPack.load_from_file('http://example.com/',
+                                     'tests/testfiles/yaml_inclusion/2021/01/20210101.yaml',
+                                     TagPackSchema(),
+                                     taxonomies,
+                                     'tests/testfiles/yaml_inclusion')
+    assert isinstance(tagpack, TagPack)
+    assert 'title' in tagpack.schema.header_fields
+    assert tagpack.contents['title'] == 'BadHack TagPack'
+    assert tagpack.contents['abuse'] == 'scam'
+    assert tagpack.tags[0].contents['address'] == 'bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh'
+    assert tagpack.tags[0].contents['context'] == '{"validated": true}'
+
+
+def test_yaml_inclusion_overwrite_abuse(taxonomies):
+    tagpack = TagPack.load_from_file('http://example.com/',
+                                     'tests/testfiles/yaml_inclusion/2021/02/20210201.yaml',
+                                     TagPackSchema(),
+                                     taxonomies,
+                                     'tests/testfiles/yaml_inclusion')
+    assert isinstance(tagpack, TagPack)
+    assert tagpack.contents['title'] == 'BadHack TagPack'
+    assert tagpack.tags[0].contents['address'] == '1Ai52Uw6usjhpcDrwSmkUvjuqLpcznUuyF'
+    assert tagpack.tags[0].contents['abuse'] == 'sextortion'
+
+
+def test_empty_tag_list_raises():
+    tagpack = TagPack.load_from_file('http://example.com/packs',
+                                     'tests/testfiles/simple/empty_tag_list.yaml',
+                                     TagPackSchema(),
+                                     taxonomies)
+
+    with pytest.raises(ValidationError) as e:
+        tagpack.validate()
+
+
+def test_multiple_tags_for_one_address_work():
+    tagpack = TagPack.load_from_file('http://example.com/packs',
+                                     'tests/testfiles/simple/multiple_tags_for_address.yaml',
+                                     TagPackSchema(),
+                                     taxonomies)
+    assert len(tagpack.tags) == 2
+    tagpack.validate()
+
+
+def test_duplicate_does_not_raise_only_inform(capsys):
+    tagpack = TagPack.load_from_file('http://example.com/packs',
+                                     'tests/testfiles/simple/duplicate_tag.yaml',
+                                     TagPackSchema(),
+                                     taxonomies)
+
+    tagpack.validate()
+    captured = capsys.readouterr()
+
+    assert "1 duplicate(s) found" in captured.out
+    assert len(tagpack.get_unique_tags()) == 1